backend/requirements.txt
fastapi>=0.100.0 # Framework
uvicorn>=0.22.0 # Server
pydantic>=2.0.0 # Data validation
sqlalchemy>=2.0.0 # Database ORM
psycopg2-binary>=2.9.5 # PostgreSQL
python-dotenv>=1.0.0 # Environment variables
requests>=2.30.0 # HTTP requests
nltk>=3.8.1 # Natural language processing
alembic>=1.11.1  # Database migrations
pydantic-settings>=2.0.0  # Settings management
textblob>=0.17.1  # Text analysis
spacy>=3.5.0  # Natural language processing